Refactoring + adapt runner code
