+++ src/middleware/boilerplates/runner/index.js
@@ -4,7 +4,6 @@ const { ServiceBroker } = require('moleculer');
 const os = require('os');
 
 const createServices = require('./createServices');
-const configureExpress = require('./configureExpress');
 const CONFIG = require('./config');
 
 // Broker init
@@ -16,7 +15,6 @@ const broker = new ServiceBroker({
 });
 
 createServices(broker);
-const app = configureExpress(broker);
 
 // Start
 broker
@@ -28,15 +26,4 @@ broker
   )
   .then(() => {
     console.log('Server started. Node ID: ', broker.nodeID, ' Transporter:', broker.transporter, ' PID:', process.pid);
-
-    app.listen(3000, err => {
-      if (err) {
-        console.error(err);
-      } else {
-        console.log('Listening on port 3000');
-
-        // Start REPL mode, which allows to send commands through the CLI
-        broker.repl();
-      }
-    });
   });
