+++ src/middleware/boilerplates/runner/package.json
@@ -18,20 +18,11 @@
     "@semapps/sparql-endpoint": "^0.1.0",
     "@semapps/mime-types": "^0.1.0",
     "@semapps/webhooks": "^0.1.0",
-    "cors": "^2.8.5",
     "dotenv-flow": "^3.1.0",
-    "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "jsonwebtoken": "^8.5.1",
     "moleculer": "^0.14.3",
     "moleculer-web": "^0.9.0",
     "moleculer-repl": "^0.6.3",
-    "moleculer-db-adapter-mongo": "^0.4.7",
-    "node-fetch": "^2.6.0",
-    "node-jose": "^1.1.3",
-    "openid-client": "^3.11.0",
-    "passport": "^0.4.1",
-    "passport-cas2": "0.0.11"
+    "moleculer-db-adapter-mongo": "^0.4.7"
   },
   "devDependencies": {
     "nodemon": "^2.0.1"
