+++ src/middleware/packages/connector/package.json
@@ -9,10 +9,12 @@
     "@semapps/triplestore": "^0.1.3",
     "base64url": "^3.0.1",
     "jsonwebtoken": "^8.5.1",
+    "express-session": "^1.17.0",
     "moleculer": "^0.14.3",
     "moleculer-web": "^0.9.0",
     "node-jose": "^1.1.3",
     "openid-client": "^3.11.0",
+    "passport": "^0.4.1",
     "passport-cas2": "0.0.11"
   },
   "publishConfig": {
